refactor(perl-lexer): split checkpoint module and add edge tests (#0000)
